pyproject.toml
[tool.poetry]
name = "modules"
version = "0.1.0"
description = "A project exploring the integration of LLMs with UI testing"
authors = ["7db9a"]
license = "MIT"
include = ["modules/*"]

[tool.poetry.dependencies]
python = "^3.8"
pyautogui = "^0.9.53"
Pillow = "^9.2.0"
python-xlib = "^0.33"
pytesseract = "^0.3.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
